ncaab -> ncaamb
